+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Removed
 - defensive NameError handling in ChangelogCLI.abort method
 ### Fixed
+- replace token {OPENCOLLECTIVE|ORG_NAME} with funding org name
 - prefer .example version of .git-hooks
 - kettle-commit-msg now runs via rubygems (not bundler) so it will work via a system gem
 - fixed logic for handling derivation of forge and funding URLs